LayoutTests/imported/w3c/web-platform-tests/pointerlock/constructor-expected.txt

Create Pointer Lock events and check each default value.


PASS Default event values for mouse event interface and its pointer lock extensions.
FAIL Default event values for pointerlockerror using a dictionary assert_equals: expected 10 but got 0
